Move cassette components into subfolder
