+++ src/components.d.ts
@@ -13,11 +13,11 @@ declare module 'vue' {
   export interface GlobalComponents {
     AlbumList: typeof import('./components/AlbumList.vue')['default']
     AppFooter: typeof import('./components/AppFooter.vue')['default']
-    Cassette: typeof import('./components/Cassette.vue')['default']
+    Cassette: typeof import('./components/cassette/Cassette.vue')['default']
     CassetteActionsBar: typeof import('./components/CassetteActionsBar.vue')['default']
     CassetteInfoDrawer: typeof import('./components/CassetteInfoDrawer.vue')['default']
-    CassetteItem: typeof import('./components/CassetteItem.vue')['default']
-    CassetteSide: typeof import('./components/CassetteSide.vue')['default']
+    CassetteItem: typeof import('./components/cassette/CassetteItem.vue')['default']
+    CassetteSide: typeof import('./components/cassette/CassetteSide.vue')['default']
     PlaylistList: typeof import('./components/PlaylistList.vue')['default']
     RouterLink: typeof import('vue-router')['RouterLink']
     RouterView: typeof import('vue-router')['RouterView']
