+++ src/components/cassette/CassetteSide.vue
